+++ docs/boksa.md
@@ -1,116 +1,209 @@
 # Body of Knowledge, Skills and Attitude
 
-We beschrijven de BOKSA voor FDND aan de hand van de DLC. Dit is een work-in-progress en verre van af.
+We beschrijven de BOKSA voor FDND aan de hand van de DLC. Dit is een work-in-progress en verre van af, vooralsnog staat hier een opsomming van elementen die we tijdens de DLC uitvoeren.
 
 ## Analyseren
 
 Project Board
+
 Github Projects
+
 Trello
+
 Linear.app
+
 Mindmap
+
 MoSCoW
+
 Customer Journey
+
 Debriefing
+
 Definition of Done
+
 Briefing
+
 Gesprek opdrachtgever
+
 TeamCanvas
+
 Todo List
+
 Contentanalyse
+
 Persona
+
 Code Conventies
+
 User Stories
+
 Job Stories
+
 Planning Poker
+
 Performance Budget
+
 A11y Budget
 
 ## Ontwerpen
 
 Data Model
+
 Harris Profile
+
 Breakdown Schets
+
 Breakdown Chart
+
 Wireflow
+
 Styletile
+
 Atomic Design
+
 Moodboard
+
 Living Styleguide
+
 Schetsen
+
 Figma
+
 UI-Stack
+
 Morphological Chart
+
 Wireframes Wireflows
+
 Divergeren Convergeren
+
 Sitemap (Hierarchie)
+
 Kleuren en contrast
+
 Pen en papier
+
 Adobe XD
+
 Schetsen
+
 Responsive Design
 
 ## Bouwen
 
 HTML/CSS/JS
+
 Living Styleguide
+
 Serverside Rendering
+
 Clientside Rendering
+
 Progressive Enhancement
+
 Component Library
+
 Libraries: Chakra, Tailwind, Three.js, Threlte, Storybook, GSAP, Framer, socket.io
-Tools: ESLint, Prettier, Typescript
+
+Plugins: ESLint, Prettier, Typescript
+
 BEM
+
 Editors, Browsers, API's
+
 NPM (Yarn)
+
 Frameworks: Svelte, React, Vue, Angular, Express, Gatsby, 11ty, Ember, Hugo
+
 Semantiek
+
 Afhandelen van issues
+
 Mobile first
+
 CRUD
 
 ## Integreren
 
 Github Pages
+
 (S)FTP
+
 Pull-request
+
 Activity
+
 .env
+
 wiki
+
 docs
+
 Readme.md
+
 Gitbook
+
 Vercel
+
 Adaptable.io
+
 Railway.app
+
 Fly.io
+
 Github Actions
+
 Netlify
+
 Git CLI
+
 Google Cloud
+
 NPM
 
 ## Testen
 
 Tab/Keyboard test
+
 Kleurenblindheid test
+
 Brillentest
+
 Feedback sessie opdrachtgever (acceptatietest)
+
 Usertest
+
 Testen in Usability Lab
+
 Colorcontrast
+
 Browsertest
+
 Elderly test
+
 Performance test
+
 Waterval performance
+
 Expert review
+
 Client review
+
 Code review
+
 Design review
+
 lighthouse
+
 validator
+
 WCAG-test
+
 a11y
+
 screenreader test
+
 device test
 
+Quality assurance
