+++ src/helpers/markdown-parser.js
@@ -22,15 +22,15 @@ export default unified()
   .use(remarkGfm)
   .use(remarkRehype, { allowDangerousHtml: true, footnoteLabel: 'Noten en citaties' })
   .use(rehypeRaw)
-  .use(rehypeShiftHeading, { shift: 1 })
+  .use(rehypeShiftHeading, { shift: 1 }) // Shift headings up, only h1 in header, h2 in documents
   .use(rehypeSlug)
   .use(fdndDiscussions)
   .use(rehypeWrap, { wrapper: 'article' })
-  .use(rehypeToc)
+  .use(rehypeToc, { headings: ['h3', 'h4'], cssClasses: { toc: 'subnav' } })
   .use(rehypeAutolinkHeadings, { properties: { class: 'auto-link', 'aria-hidden': 'true' } })
   .use(rehypeSectionHeadings, { sectionDataAttribute: 'data-heading-id' })
   .use(shiki, { theme: 'monokai' })
   .use(fdndWrapper)
-  .use(rehypeDocument, document)
+  .use(rehypeDocument, document) // er moet nog tabindex="-1" op <body> om het menu mobiel te laten werken
   .use(rehypeFormat)
   .use(rehypeStringify)
